+++ src/tables/mod.rs
@@ -7,6 +7,7 @@
 
 pub mod cmap;
 pub mod context;
+pub mod device;
 pub mod gdef;
 pub mod gpos;
 pub mod gsub;